Patches for pdroid aosp-mod
This is open for discussion, but it works, with three positive tests. Many
builders of AOSP kangs cherry-pick the framework2.jar commits, which
is causing users of those patches to fail the boot.img patch. There
are several other ways to do this, but modifying the script to accomodate
this seemed ill-advised. My second-best idea is to somehow do this same process
through special_instructions.sh.