Move provisionals to platform specific folder
This is being done in order to make way for a CM7 provisional. But really this
is where we should go for a bunch of stuff. Too many of our platform specific
actions are global now. And things are in a "common" directory when they are
only common to JB. This will also clean up some of our more confusing naming
practices, once we finish the transition.

I only am doing this small step so you can look at it first. Also, moving the
entire "common" directory would break all our symlinks and take a bunch of
work, lol. Baby steps.