Move files to be copied into $P/$R/$B
This is being done to accomodate the new framework2.jar that is coming
out tomorrow. It should really only affect the pdroid aosp-jb patches.
But this is part of a larger movement to move things to place where they
more naturally belong. If certain files/functions are only needed on
certain build dates, then that is logically where those things should be
stored. Of course, symlinks are deployed in order to not hold multiple
instances of the same object.

The common folder should actually be moved into 'jellybean', but I have
renamed enough symlinks for today. Moving the framework.jar.sh into
$P/$B/$R is next commit.