Fix /r character probelms in getbuildprop function
Suggested by bundik on post #957. It has been responsible, I guess,
for the excess RIL deletion that sometimes causes FC's. I have checked
it and it runs, but since I have never had the problem, I can't comment
as to that. Let me know if it looks OK to you.

The rest of these changes are moving the framework2.jar.sh into
$P/$B/$R as promised. I know it seems redundant. But since that
script is really only needed in certain build dates, I believe that is
where it should be stored.

I am thinking about leaving the possibility open for 'special_instruction'
type scripts to be ran at all three major stages of the script. This would
be the second stage. As it stands, it is a pdroid-specific script just
sitting in the middle of the AP proper.