Advanced build date query- Do Not Merge
This is not ready. It works AWESOME on single mod expressions. But no
matter how I have tried to intergrate it later in the script, when more
than one MODTYPE is called, it then defaults to latest for both (which
is pretty much what happens now) but worse, only the second argument's
patches get applied.

It is a super-cool addition to the program (I think) and would probably
cut our error reports in half. I am pushing it to 2.1 so that maybe pastime
could take a look at it this weekend as well.

No pressure, though. I will fix it if I need to. Cool though, huh?
Based the build date off of UTC.