Support for extension; AOSP-MOD
Still going to be a bunch of failed hunks, I haven't had wbedard build
specifically for this. I did that for PDroid, all that needs to be done
is make a stock build with the framework2.jar classes already moved.

It is cosmetic, but I just now had another user posting about ignored
hunks. It makes them nervous.