Broken patches for CM7- DONT MERGE
These don't work any better than the existing latest patches. In every case,
the roms will not boot past bootloader, even when applied to the stock rom
the diff was made from. When I flash either the pdroid-patched build or the
stock build, both boot. So there is some sort of serious problem with the
process itself here.