Build Date Query final
This works perfectly for what we have now- not done in the long-
term sense.
* If one mod is passed, it finds the best patches for it.
* If a date is specified, that date is used instead of the query.
* If multiple mods are passed with no date- all default to latest
* Multiple mods and one date- pdroid uses the date, rest default latest

Functionally, this is all we need right now. But the best thing to do
is to tie the BID as an array, and have them used in conjunction with
PID by position. But I need to learn more before I can do that.