Major Rework of patches
With the Adv.BuildQuery, this should make it so users ONLY have to
enter the mod-types and possibly a SIMPLE romtype, i.e: cm, pa, aokp
or aosp. No more multiple types of aokp or aosp.

This is in preparation for 2.2. The only thing that really could be better
is an array of patch dates. But see the patch date commit, which I will
reference in an issue so we can easily talk about the changes.