Version bump 2.2.1
Hopefully everything works, I am going to give this an extra tag. Probably
as fully tested as anything we've put out, by the use of ApG while in
progress. There are some little bugs but nothing that impacts functionality.
The only open question is whether the script fixes the build query for OSX.
They can override with manual date input, though. But if the case function
correctly leads to the OSX BDID, then the query should work.