Bump to 2.2.15
Hoping for a 2.3 stable for next official update. If the OSX stuff is
truly fixed (no reports back yet) that is the last MUST-DO that I know about.

Then, hopefully, this update will be complete and we can go back to normal
for awhile.

That will be when the pdroid and pd2.0 patches break again.