New '-h' menu, compatible with OSX
Not really my favorite, but the OSX ls doesn't have an ignore. It was
really a chore getting the output to look like I want. The real answer
is to probably filter it through a perl or awk script, but I didn't want
to add a perl requriement, and I will look at awk later.