Bump to 2.3.0
The idea is for this to be the stable release for awhile. I am going
to give it to the ApG/git users, and if nothing arises, a tagged&bagged
version will go on techerrata late tonight/early tmrw morning.

pastime, if you spot anything when you get this chance to look it over
we can talk about fixes then.