Add explicit message to $LOG explaining provisional copy
We are getting a bunch of confusion about this; mostly related to the fact
that ApG is configured to answer the 'yes'/'*' question w/o the user
knowing that it was asked. Since at this point many users are conditioned
to quote from the log, I am hoping this helps.