Bump to 2.3.1
Necessary for older roms so official release is going out. Improve build
query logic.

Beginning of 'Operation Clear', where I intend to begin
commenting extensively, as well as other insane things like rename variables
to be more descriptive.

Bound to be a REAL hassle, but I think it will inspire
more community involvment if the script is easier to follow.