Support for Evervolv
And a rearranging of some of the other patches. I had been operating on the
assumption that our commit to relieve the method cap had gone to all
the variants, which has turned out not to be the case. AOKP and Evervolv
have not merged that change, for some reason. Frustrating.

So I have re-added the framework.jar.sh to some patches, as well as added
some patch sets. This will unfortunately probably take an update or two to
get completely correct.