Add support for $RID-specific provisionals
Just tests to see if there is an RID specific file before moving on to
use the common provisional. This actually makes us a lot more flexible. When
we go to remote repos, this will probably allow us to patch any rom we want,
since we can host a patched+decompiled version of every jar to chery-pick
from.

Big implications from such a little change.