Error catch for unexpected method cap issues
A bunch of random forks have, for some reason, not merged the commit
we submitted to CM that fixes the method cap issue. I am instituting this
so we don't have to keep messing with people's RILs unneccessarily but
are still able to solve this problem.

The original framework.jar.sh call is being left as is, so that roms we
know need it can get it without having to deal with compiling that jar twice
or thinking there are errors on our part when there are not.