Rework installation of framework resources
It was reinstalling them for each file in AID[@]. This will also install
the resource table if APKTOOL is used at all, even if framework-res.apk
is not in files.txt. This is needed for stock roms, which are in fact
probably the only roms that need the resource table at all.

Better practice this way...and a faster script.