Adjust build query for Cm7
Defaults to latest since CM7 does not have UTC time stamp. I just realized I
should probably change this to test for if $UTC is empty or zero. But I need
to get this out. I will make an issue to revisit it later.

Finally, this plus the $API fixes the issues with CM7 we've had since the
first day of release.