Test run at 4.2x ParanoidAndroid error-catch
I am not sold that this is the answer...especially because I wasn't able to
get any traction on getting my error messages into ApG. But I don't want pa
to be all I hear about, but I also don't want to mess with the ROMTYPES list
anymore. Not until I do away with it for good.