Add the ability to use the smali binary on apks
To be used for files that apktool chokes on. We really only NEED apktool
for framework-res.apk, I am pretty sure. I think all the res and src can be
gotten with smali.

The way I implemented this was nice and simple. And tailor-made to give
pastime an aneurysm, if he was still with us. This is the kind of thing that
drove him mad. I try to limit it...