Add apk_was_copied variable (Incomplete)
I finally saw that the apk, if a prepatched version was copied over due to failed patches, was being rebuilt
by apktool, and concievably overwriting the prepatched version. I hope that this will perhaps influence
the disappearing Mms.apk bug, but it at worst speeds up the auto_patcher significantly in these cases.

It is incomplete in that at this point no patch sets that both copy over prepatched files AND have more than
one apk are supported. None of the current patch sets match this profile, but I try and keep the program
flexible. It needs to be reviewed.