implement apk_was_copied and APKID to fix apk handling
This uses $APKID as a working copy of $AID, all the apks. This better
handles the times when a prepatched apk is copied as a binary...this removes
the apk from the rebuild list while leaving it on the master list to be
included in the update.zip.

The $apk_was_copied allows us to also copy over an apk if patching succeeds but
apktool chokes out on us. It is set when an apk is copied and then skips the
rebuild line, avoiding the overwriting scenario we had before.