AUTO_UPDATING feature, .config implementation, and versioning for patches.tgz
Maintainers and Contributors!
 :There is also a new variable PATCHES_VERSION, which must be incremented EVERY
time there is changes to ANY patches. Since we sometimes edit older patches,
I intend this variable to be set as equivalent to the date of the edit.

This commit sets up a .config file, with two current values
  AUTO_UPDATE=true
  PROUCTION=true
PRODUCTION is currently unused, I have it in mind for a test suite.
AUTO_UPDATE is fully implemented for Darwin/Linux. Cygwin/ApG users can test
this feature by uncommenting line ~1005, AUTO_UPDATE=false.

This feature checks if the user's branch is behind origin/master, and if so
merges that branch, if it can communicate with github (network access). The
script is then ran again, checking to see if batch.sh is needed to integrate
new patches and if so, running it.

Users who edit their own patches/script will find this feature troublesome.
It can be disabled by setting the AUTO_UPDATE key to false in the .config file.

There is a new temp file, .updated.log, which holds the merge and update info
to be added to the final logfile. Neither this nor the .config file is under
version control by design.

If the git pull fails (auto_update_error) I am unhappy that the logfile is so s$
of the rom probe information to run in this case, but ONLY if fails...I don't w$
twice in case of successful updates, etc.