Set LANG to C to optimize grep, search, and sort.
This likely speeds up patch as well, although I have not benchmarked it personally. I have seen that some
scripts can run 3-5x faster with the proper locale set and thereby skipping the encoding. Since this script
relies heavily on text search and replace, I am guessing this will make a big difference, while still
paling when compared to the decompile/recompile time.

This implementation is somewhat non-standard in that it has the courtesy to not permanently clobber the
user's global locale settings.

Close issue #104