Add last commit to LOG (reduces need for such incessant versioning)
Since we are basically requiring everyone to have the .git folder, I am going
to start leveraging it in some fashion. This goes hand-in-hand with the
auto-updating...without this, I would have to up the version each commit.

Versioning will still be very regular, and completely non-sensical. Hopefully
I can be more disciplined about that post-3.0