Fix insidious little bug that would create a false breakage in APK patching
If multiple apks needed to be patched and the first apk had a failure that
we resolved another way (copying binary or PROVISIONAL use) the fact that we
concatenate the APK.logs would have grep out the first failure for all
subsequent apks. I don't know if I have seen this, but we have cases in the
project that could trigger it now, I believe. A broken Mms.apk in OPD combined
with tabletUI would report a broken tabletUI, for instance.