Redirect stderr to the LOG in a couple more places
In my checks for the ALTERNATE_PATCH feature, I encountered (some possibly
irreplicatable in the wild) situations where cp could not find the files
it was supposed to copy. But the err went only to output, so we would never
know by looking at the LOG. This should be added to a function, maybe, and
checked if needed elsewhere.