Fix CONFLICTS function
When using rom probe, the CONFLICTS function did not iterate through the mod
list. So if there were multiple mods and no rom type specified, the conflicts
function would only work on the first mod in the list, generally hard-coded
to be a PDroid mod.

This updated version of the fix resets the PID variable, so each call to the "mods_info"
function sets the various mod-related variables as set to each individual mod. The
PID array (which is the list of requested mods set at the CLI) is afterwards reset to the
ordered list made previously (SPID array).

Someday all of these various variables and arrays need to be sensibly renamed.

Closes (#125)