Revert "Fix CONFLICTS function"
Whoops. Needs work....if used with a pdroid mod, it messes up and tries to
apply one of the mods twice. Probably conflicts with the function that ensures
the pdroid-type mod is the first applied (which we do as a hack to make sure
that using prepatched "provisional" files doesn't overwrite other mods...if
pdroid is applied first then we cam layer the other mods on afterwards.
This reverts commit 71b9ade837d8bd269e7a7a96d4a4af171630ca33.