Add a git check to auto_update, fix some OSX incompatible code and rearrange function that works, but shouldn't
That check for the git directory was supposed to be testing true...it was only passing b/c it was calle
from a sub directory. The removal of &>> was b/c OSX reports it errored. More testing on OSX is needed
but I haven't turned off auto-updating by default for then quite yet