Move manifest and certificates from orig Apk into the prepatched copy (if used)
This should help with the disappearing Apk, if this implements it correctly. I
still think I have seen the disappearances even with succesful patching though
which preserves the rom's certs/manifest by default. so this will not be a
general solution unless I am missing something. I am trying to avoid having
to try and sign all the apks, there are security ramifications of that and I
am not really willing to manage all the various key juggling it could entail.