Hide the "tpu not found" errors in Cygwin
Cygwin is not a full unix environment, and does not always have access to all
the utilities by default. Turns out tput is one of those. Put I am just
using that to output some color to the successful patching, trying to cut
down on the erroneous bug reports. This redirects to stderr.