Manual date specification on CLI was failing
This was caused by removing the $P variable yesterday. We need to rewrite our
variable names, as being haphazrd with them has left us with problems of scope.
Some of these problems are very hard to find because they have variable set
elsewhere and we don't even know it.

Previously, ./auto_patcher encore.zip openpdroid cm 20130516 was failing
because the script was checking for [ -d patches/$P/$ANDR/$RID/$BID ]. But
$P was no longer defined, because I changed it yesterday. But we did not know
that was where we had set $P for this function, that was an accident. This is
how the most frustrating bugs come to be, scope and variable confusion.