Close big bug in Incremental function
Incremental patches are in place as a method to deliver hotfixes. If we have
some reasonably small change we wanted to make to a big mod like openpdroid, we
would build one patched OPD, and then one patched + the bugfix. The resulting
diff becomes a single feature patch. You can see this at work with
services-gp-fix.patch in Openpdroid.

However, I just saw that there is a big bug. We apply patches in the program,
and send the output to a temp file (JAR.log). We then scan that file for a
FAILED string, and if found it triggers our provisional or returns a failed
patch and exits. When applying the original INCREMENTAL patches, we were
overwriting the JAR.log and all alerts of a FAILED patch.

Simply changing the INCREMENTAL patches from ">" (overwrite) to ">>" (append)
solves this issue.

Big bug, though. Generally speaking, if this had been causing problems, we
would have already known because services.jar would have refused to compile
(which is how I found out about it) or we would have reports of boot loops. But
eventually, the upstream services.jar would have changed to the point we would
have gotten reports. It would have been hard to track a boot loop, glad it is
found now.