Add a .gitattribute to treat patches and text files as binary (from gcydtmkq)
I had never heard of this, but it is a great idea. This will treat our patches,
at least, as binary files and hopefully preserve our line endings. We are
starting to see some problems in Cygwin when not using ApG...I don't know if
that is because of a change in Cygwin or git. But attention needs to be paid.

Hopefully this has an effect. I also tried to add our patches/* path, since
all of the files in there are processed. We will see if it can be used that
way. I could also have made this untracked, but I want contributors to be able
to post patches, and this stops them from having to set their own unique
environment.