New services2.patch for PA
This is another use of the ALTERNATE_PATCH structure. PA has multiple forks
floating around, but two main forks, which are indistinguishable from the
build.prop. We use build.prop parsing to match patches, by romtype and date.
But the PA forks have identical build.props.

We previously got around this with the telphony patch by adding an alternate
patch function. If a $FILE.patch fails, it looks for a $FILE2.patch. If that
fails, it will look for a $FILE3 and so on. Might be how we leverage the change
to 3.0, maybe not. But it works for PA right now.

I also had to move the ALTERNATE_PATCH_NUMBER lower in the loop, so it would\
intialize for each $FILE being patched. Where it was, it carried the increment
over to the next patch, as in started looking for telephony-common3.patch.