Remove patch check for cygwin
I need some sort of institutional way to handle the dependency messages. Right
now it still while tell Cygwin users that they need patch. But that is a
minimal cost, since we would override any patch install with our binary.

But still, a better way to handle all that would be dependencies in OS-specific
arrays. For output and checking.