Bundle patch.exe.manifest (thanks to @thewaronerror)
Patch is blacklisted when under UAC. That is why we are getting permission
denied when run under non-administator. At least this is a quirk of Windows
and not indicative of a larger problem.

Several solutions were considered, including pulling the manifest from the
cygwin packages (not ideal because it isn't under auto-patcher source control)
to embedding the manifest in the binary (since I got the binary pre-compiled,
I would rather not inject files) so this is the solution we went with. Thanks
to @thewaronerror who brought all the technical backstory and solutions.