Tie auto-updating to physical files in addition to the enviromental variables
Cygwin user reports auto-update is working, but batch.sh not reliable. I tried
to add checks for physical files we already create to replicate the sourced
variable approach that works for Linux.

I attempted to tack this on in a way that the existing implementation will
still work. This new, file-dependent implementation should actually stand
on its own...we could remove the sourced variable method should this work
as we want it to.

Addresses half of #141