Add generic folder to opd/4.3
When build.prop query (and certain other operations, search for 'generic')
fail, the patcher defaults to 'generic' and 'latest' patches. Here they did
not exist, so there was no files being adjusted. It is a bit of a mystery
why nothing got decompiled (it looks like it did not read the "files.txt") but
this should help.

The 'generic' folder is a symlink to 'aosp'. So anything the build.prop query
cannot resolve will default to aosp patches. That will have it fail loudly since
I expect AOSP patches to fail for PAC and PA (I have seen the build query fail
with both in this manner).

Try again with PAC, okay @Urinophoria?