When multiple mods selected, all mods get matched through build query
This has been a big white whale, and the fix is so small.

This sends each individual mod through the build query one at a time, and
now matches each mod to the patch date most fitting for the rom's build date.
I had tried a couple things, but I was worried that Cygwin's bash wouldn't
support it. This should be fine, and looks to work great.