Update "provisionals" for 4.3
There were some problems patching a few mako builds (in CdmaService again...)
so I updated the common provisionals we have had success with during 4.1 and
4.2.

These are smali files from STOCK builds, but builds from which we have made
patches. Since they are the source of the diff, we know that they will patch.
The result of using them is that whatever commit difference their rom has
with the source are backed out. That has the potential for a bad experience
so is best used sparingly. Generally it has been found to be okay with these
files, sometimes it is merely the register count (a smali decompiling protocol
not substantive)