Reenable the ability to specify PATCH_DATE manually (if multiple mods, only works for Pdroid-related mod)
The recent build-query fix overwrote this ability. I would like to add support where all patches could be
manually specified on the command-line, possiby seperated by commas like MODS. But there is a wrinkle. B/C of our "provisional"
usage (which copies over smali files) we have to ensure that the MOD that uses provisionals (only PDroid) is the first applied
because otherwise the danger exists of overwriting any earlier patches.

E.G. If we patch ActivityManager.smali for v6supercharger, and immediately afterwards copy over that same file as a provisional
for OpD, well then that update.zip will no longer work for v6.

Anyway, to add supportfor specifiying patch dates for multiple mods, the user would potentially be quite frustrated if they did not
know that the Pdroid-related mod had to be the first entered. A mechanism to take care of that for them would probably be best,
but would be very very complicated for something that is so unlikely to be used. Also, if we ever get to on-board patching, then
those very rare cases could just be handled by consecutive patches. So that may never happen, is the upshot of all this.