Add 'latest' folder for aosp-ever
I am not even sure we are ever using the 'latest' folders as of the finalization
of the build query fixes. But better to safe, since there are several error
catches where it defaults to latest if it cannot find what it is looking for.

Great job on all these patches...learning how to do this and github in this
amount of time is no small thing. I had some very good luck when you decided
to join the project.