Whoops. We forgot to version up.
Versioning is pretty important for us in particular. Since the users get our
changes immediately, we need to be on the ball so we know exactly what changes
they have when they submit logs. If we just pushed a fix, we need to know
if the broken log they submit has the latest commits or not. I added a logline
that lists the latest commit but the version number is easier to keep track of.

The PATCHES_VERSION needs to be incremented every time you make a change in
either the patches_tools or patches directory. That is what determines if
batch.sh runs on the users machines. If they update to get evervolv support
but we haven;t incremented the PATCHES_VERSION number, then the user
will report that it isn;t working, cause it won't be.

You can usually just do these steps at the same time as adding the patches.
For now, try to only increment the 100th place, the second significant digit.
I will increment the first digit and the changelog when necessary. Thanks!