Change CM10 files.txt to reflect the baksmali use for Mms.apk
Files.txt is one of the most important files in the program...it tells the auto-patcher
which files to operate on and which tools to use. When the Mms.apk was giving me fits
(disappearing on everyone) I finally switched to using baksmali instead of apktool
that problem went away. As I was just testing the change, I only added it Cm10-1.
I found out it worked, but never ported it down to 4.1. The broken Mms was the result.

I also added a provisional I found useful. I intend to start a page on thewiki where we
build a knowledge-base of useable/not-usable provisional files. It will be MOD-specific,
but might give contributors an idea of where they can possibly fudge.