Add framework2 to cm patches to support temasek builds (3.0.212)
I am almost to a prototype of what will make this method unneccessary. But
in short, if a $NAME.patch fails, the patcher will look for a $NAME2.patch and
try to patch using that.

I call it an ALTERNATE, and that is just what it is, a fallback patch to
serve when we know a subsect of roms fail reliably and we cannot use
provisionals to fix it. I only use it for real popular roms, or in this case
for @temasek, who besides being popular, has long been a supporter of the
project and has exposed many many of his users to the program.

Thanks to @DualJoe for all the input and direction