Beginning of Autopatcher 3.0 implementation
Not working. Still lots of bugs. patch_matcher function has not been added yet.
The INCREMENTAL needs to be moved out of PATCH_QUEUE (due to use of --dry-run.)

But this does work insofar as it takes the *smali patches and uses them.

Big changes coming, so commit seemed prudent.