Add new hidden folder, the ".reservoir"
I agonized over this, I really did. It doesn't quite fit our domain-model.
But I couldn't think of anything easier, considering that we know ApG is in
its final form and must be worked around.

This model may change, but the reservoir is going to take the place of the
ALTERNATE patches which are now deprecated. Basically, I intend to offer support
for any open-source rom. If the Autopatcher fails to patch, then the user just
needs to ask the ROM developer for a link to his manifest and post their log.

Then, I or Urinophoria can build the module (so if services fails to patch, just
build the services with 'make services', takes maybe 15 minutes). The smali
patch we need to add can then be split out using the split_patch_opd.py and
tossed here into the 'reservoir'.

Other options include just adding them under .aosp, so if it is...BubbleVenom
ROM, then .aosp-bubblevenom might be fine. But I intend the reservoir to be
a place to stash needed smali patches that don't fit anywhere else. If you
need/want to add a file to the reservoir, you can simply add a new date. None
of the params inside the reservoir matter (too much). For now, we only have
patches/openpdroid/4.2/.reservoir/pa/20130703, but the 'pa' portion is just
for our information. As long as the patch keeps the name format (auto-generated
by split_patch_opd.py), it can be found by the new model.