Version up to 3.1
This is for testing fellas. Let me know how your runs go, if you have time
to check it out. The script is run exactly the same for now, but you should
find that your rom (assuming it patched before) now patches no matter which
ROMTYPE you specify on the command line.

Essentially, we now use SMALI.patches instead of JAR.patches. When a SMALI.patch
fails, the program now searches for other possible matches and tries those as
well. It shouldn't take too much longer to run that it did before, either.

I experimented with matching the files by a content-based system, but it seems
that using the name is enough for now. You will see I had to use some new
processing scripts for the patches, I will prepare a document with some info
on them later. @Caldair, this process has not been ported to APKs yet, but
that should be pretty easy. Tae it for a spin if you can, and we will
then take a look at working it for tabletUI.

I have a bunch of commits here, and it would be pretty hard to see what I was
doing from commit to commit. Here is a diff of the changes squashed. This
is what will probably end up in the main branch when it gets merged.

http://pastebin.com/PRiYu4K0

Criticism and ideas/feedback welcome!