Remove all front-facing ROMTYPE messages
They are misleading at best...we now will be using a hodgepodge of patches
best to retrain and get people used to the simpler methods.

I cannot remove it completely, due to ApG, and it DOES have some benefits for
us as developers and testers. I would prefer to make it available only to
PRODUCTION=false (in .config, currently unused) environments. But it is
locked in as a param as long as we need/use ApG.