Update some of the help text
After this goes public, I hope to be available to build patches for
basically any fail. The resistance before was that the effort was large and
could end up only helping one person, one week. But now that all patches are
available to be added to the patch library, things change.

My observation/contention is that there are only so many changes to these
low-level files being passed around, and that just about every rom is just a
hodge-podge of theses changes. So I believe that once we build up the library
just about every AOSP-based rom will fall in line.

We will see if I live to regret this. I hope to open up either a github issue
or a Google form, where I will be asking for LOG file, link to rom and link to
rom source code...with perhaps a policy of only supporting rom's with manifests
available. I certainly don't intend to debug build errors for every random ROM.