Change framework2 (for temasek) from ALTERNATE to '.reservoir'
This is a perfect example of what the reservoir is for. This patch is just for
temasek builds. So instead of starting .cm-temasek, or whatever, we just throw
these patches into the '.reservoir' and hope that they cover us for the
rest of 4.3. As long as there is only this one difference between temasek and
CM, this should be fine.

There is the case where CM happens to modify that file in a different way, then
we may have to rebuild for each. But I bet that is very rare. Incrementals
could also be applied, although that imterface needs work.