Capture stderr on auto_patcher script and fail if stderr occurs
We have sometimes fallen victim to silent stderr, where stderr is only printed
to screen. This is very troublesome, because we don't log it, and ApG users
would never even see it.

This is an extreme solution, right now the script fails if ANY unhandled errors
occur. We also now log them, so at least we should be able to see what happened.
Time will tell if I need to relax the complete failure outcome.

Essentially, the exec command redirects all stderr to a file, and at the end
of our major operations, we make sure the file is empty, or we halt operations
and display an error message.