Change up globbing on context_editor
This is supposed to grab all the ContextImpl patches in a given directory
and add them to the PATCH_QUEUE. I was given a log in a situation where the
path was expanding to the correct spot, but wasn't "globbing" (picking up
the files with the wildcards) the actual files. I hope that this somewhat
more straight-forward approach will prove a little more portable.

The original was a remnant of the period when I was trying to handle different
ContextImpl files differently and had to be very careful of the random "$"
characters that are inserted in the file paths by the decompiler. Since we
now grab all ContextImpl files, we can safely ignore that ACII character and
glob away.