Add CARBON support
Easy easy. I have a script that goes through all the directories and replaces
any duplicate file with a hard link. I am thinking of running that through the
patcher. There are a BUNCH of duplicate files at this point.

Add Carbon as an official ROMTYPE

The reason for this is in the function 'context_editor'. Our patches for the
10-13 ContextImpl files we patch have to be handled carefully. This is what
delayed the update for several months, finding the time to reason this out.

The last ContextImpl file creates a ContextImpl$4x file. It adds one more
subclass to the class. So it creates a file all by itself. This makes it
unique, in that a patch that creates a file can never fail. But we
wouldn't want to create four new files instead of just one, which is the
danger with the new model where we search for patches solely on the criteria
that they applu successfully.

What I did, was I counted the amount of ContextImpl files in each romtype. Turns
out, there are two main bases, CM and AOSP. CM needs ContextImpl$45 patches,
aosp needs ContextImpl$41 patches, both of which have the same purpose.

CM has 49 Context files, AOSP has 44. Turns out Carbon has 47, so it is it's
own beast. Adding a romtype fixes this, no matter what romtype is called, the
context_editor function will make sure the correct ContextImpl patches get
applied.