Begin APK support and add STRIP_LEVEL
For a long time, we have had to separate the APK and JAR codes, because
(due to a quirk in our very early process) apk patches were made from two
directories above and jars from one directory above. So the patch command for
apks was '-p2' and jars was '-p1'. Adding a strip level brings about the
ability to combine those two sections and reduce the code duplication by a
bunch.