Port 3gdongle to 4.3- add DEVICE_LIMIT for mods that support only certain devices
Similar to ROM_CONFLICT, this is just a list of supported devices. If the rom
is not for a device on this list, then it fails to patch. One of the patches
is for uevent.grouper.rc, so really this is grouper specific. I think we could
probably just add a bunch of others, but the demand is almost nil. I continue
to port it because I think it is neat, no other reason.

I also added a ROM_CONFLICTS entry for 'cm', since it looks like they already
merged this mod at one point.