Add 0831 CM patches to 3.1 branch
They upped the number of ContextImpl files! Bastards. Looks like it may have
even come from AOSP. I did not see that coming. I had to really take the
knife to the context_editor function, you will see how ugly that is next.

I need to perhaps rethink this entire method. It works, but if the ContextImpl
is going to be really variable this may not be safe. I could use the results
of the proberom exclusively, but that leaves us open to problems should it
fail. Hmmmm.