Reworked 'context_editor' function; remove dirty hacks
OpenPdroid creates a new subclass of ContextImpl, one that goes at the end
of the ContextImpl register. So if the last file in the stock rom is
ContextImpl$44, OpD creates ContextImpl$45.

This function now determines, by counting the number of ContextImpl files in
the stock rom, what the name of the ContextImpl file OpenPDroid needs to create
will be called.

It then looks in all patch sets that match the proberom results for romtype
and checks to make sure those patches create that exact new file. Applying
the wrong new file just gets ignored and fails silently, often causing boot
loops. This should fix that.

We didn't have this problem with the old implementation, because larger
problems stopped the user from getting a zip. Now we handle all those
probelms earlier, so this was a last line of defense.