Fix patch filtering for Cygwin
Bottom line is that Cygwin could not handle the piping to xargs. So I just
pulled the 'grep' from a pipe into a for loop. The patches are being filtered
in two cases...1) Checked to make sure the patches DO create a new file and
2) Check to make sure they DO NOT create a new file.

Still having a lot of trouble on the testing version of autopatcher on cygwin
because the long slaved-over bundled 'patch' utility is not working. So new
installations will not be able to use the autopatcher.