Clean up and comment; add some sanity checks and error catches
Now checks for globbing mistakes- too many matches, not enough matches and
instances where we might be attempting to match the regex chars, not the
expanded expression.

I tried to comment pretty heavily, I know it is not the most streamlined script
but I wanted anyone following along to be able to see what I was thinking.

There is another huge round of cleanup left, where I am going to take a hachet
to the unused functions and duplicated code. There is a lot of refactoring
available.