Add some fixes to Cygwin patch and adjust batch.sh
@Caldair, this is going to make using some of those processing scripts a little
different, sorry. I broke a part of the updater when I updated the batch.sh
script, so I just reverted that change for now.

The best way at the moment would be to go the directory with the whole patches
in it, and call the split_mod.sh script from there, specifying its location.