Add some debugging to the Provisional locator
I thought this had been fixed...some people are sending in logs where the
patching fails even when a PROVISIONAL is known to exist. Users have
checked and the file is present in their patches directory.

I am assuming that there is some problem with the matching mechanism. It got
a little hairy when we moved to 3.1. Prior to 3.1, all of our FAILS were
based on files that failed to patch...for 3.1, we now match which PATCHES
fail to patch. That is fine except for provisionals, which still are file
based. Perhaps that dsissonance is the thing to fix, but for now I would
settle for it to just work again.