Update PROVISIONALS to use find
It looks like some users have a patch binary that does not send the failed
hunks to a .rej file. We were parsing that output to find the failed file,
FAIL. This was causing the PROVISIONAL to fail.

I have switched to using 'find' against the MATCH_PATTERN. Cleaner this way
and it seems to work. There is less back and forth between switching from
handling files to handling patches this way. I also removed the provisional
copying from being it's own function...the original purpose of that function
was to provide helpful info- I have found that info to not be so helpful and
have obfuscated it. So I have moved the action so that it happens as
linearly as possible.