Update CM patches; weaken PROBE_RESULTS saftey check in context_editor
I was insisting that the build_prop be the only place where we could match
possible CONTEXT_IMPL sets. But in the end, that made it impossible for us to
patch roms where the developer may not have been attentive when composing the
build.prop.

This allows us to force a romtype in the context_editor, but does put us at a
very mild risk of bootloops, if someone tried to force an .aosp-slim on a CM
rom or vice-versa. I am hoping that I have trained the romtype out of the
user base- if so then only those who are explicitly advised to use romtypes
will do so and this risk will be ok.

closes #165
closes #163